requirements.txt
# Dependências básicas da API
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
python-dotenv>=1.0.0

# Dependências para templates e interface web
jinja2>=3.1.2
aiofiles>=23.2.1

# Biblioteca para processamento de documentos
# Usando alternativas mais leves sem dependências CUDA
python-docx>=1.0.0     # Para processamento de documentos DOCX
pypdf2>=3.0.0          # Para processamento de PDF
openpyxl>=3.1.2        # Para processamento de planilhas Excel
pandas>=2.0.0          # Para manipulação de dados tabulares
markdown>=3.5.0        # Para conversão para markdown

# Dependências adicionais para processamento de documentos
python-magic>=0.4.27  # Para detecção de tipos MIME
requests>=2.31.0      # Para download de documentos por URL
pillow>=10.0.0        # Para processamento de imagens
pdf2image>=1.16.3     # Para converter PDF em imagens
simplejson>=3.19.2    # Para lidar com valores NaN em JSON
